refactor(map): decouple construction trace from RMG
